Make d and δ preserve dtype so fields keep their declared types
d and delta allocated their result with lat.zeros(...), which defaults to
float, so they promoted integer forms to float64 even though both are exact
±1 integer combinations of finite differences.  That floated the worldline m
(via CoexactUpdate's m + delta(t)) and the Villain n (via ExactUpdate's
n + d(z)), which then relied on Batch coercing them back to int on storage.

Allocate the result with the input dtype instead.  Now every degree of freedom
keeps its known type through a full update: phi float, n int, m int, and v int
for finite W (float at W=∞, as the action declares).  This is the root-cause
fix behind the Batch value-based coercion added earlier.

+++ supervillain/lattice/compact.py
@@ -973,7 +973,9 @@ def d(f):
     if p == lat.D:
         return 0
 
-    result = lat.zeros(p + 1)
+    # d is an exact ±1 integer combination of finite differences, so it preserves the
+    # input dtype (an integer form stays integer).
+    result = lat.zeros(p + 1, dtype=f.dtype)
 
     for out_comp in lat.components[p + 1]:          # each (p+1)-form component
         out_idx = lat.comp_index[p + 1][out_comp]
@@ -1029,7 +1031,9 @@ def delta(f):
     if p == 0:
         return 0
 
-    result = lat.zeros(p - 1)
+    # δ is an exact ±1 integer combination of finite differences, so it preserves the
+    # input dtype (an integer form stays integer).
+    result = lat.zeros(p - 1, dtype=f.dtype)
 
     all_dirs = set(range(lat.D))
